[BUG] Miner not paying valid finalized budget if <20 active masternodes
Verifier: (`CheckBlock` --> `IsBlockPayeeValid` -->
`IsBudgetPaymentBlock`) expects a payment to a budget even with the case
of N<20 active masternodes on the network, provided that all N voted on
the finalized budget (the threshold is N-1).

Also remove the extra call to `IsBudgetPaymentBlock` directly from
`FillBlockPayee` (it is already called by
`CBudgetManager::FillBlockPayee` --> `GetPayeeAndAmount`).