[BUG] net: Consistently use GetTimeMicros() for inactivity checks
>>> backports bitcoin@99464bc38e9575ff47f8e33223b252dcea2055e3

The use of mocktime in test logic means that comparisons between
GetTime() and GetTimeMicros()/1000000 are unreliable since the former
can use mocktime values while the latter always gets the system clock;
this changes the networking code's inactivity checks to consistently
use the system clock for inactivity comparisons.

Also remove some hacks from setmocktime() that are no longer needed,
now that we're using the system clock for nLastSend and nLastRecv.